add validation and error handling

+++ src/main/java/ru/bironix/super_food/services/PersonService.java
@@ -7,6 +7,7 @@ import ru.bironix.super_food.db.dao.person.AddressDao;
 import ru.bironix.super_food.db.dao.person.PersonDao;
 import ru.bironix.super_food.db.models.person.Address;
 import ru.bironix.super_food.db.models.person.Person;
+import ru.bironix.super_food.exceptions.NotFoundSource;
 
 import javax.transaction.Transactional;
 
@@ -23,11 +24,11 @@ public class PersonService {
     final AddressDao addressDao;
     final UpdateMapper mapper;
 
-
     public Person getMe(int id) {
-        return personDao.findById(id).orElse(null);
+        return personDao.findById(id).orElseThrow(NotFoundSource::new);
     }
 
+    @Transactional
     public Person createPerson(Person person) {
         person.setId(null);
         return personDao.save(person);
@@ -35,14 +36,14 @@ public class PersonService {
 
     @Transactional
     public Person updatePerson(Person updatedPerson) {
-        var person = personDao.findById(updatedPerson.getId()).get();
+        var person = personDao.findById(updatedPerson.getId()).orElseThrow(NotFoundSource::new);
         mapper.map(updatedPerson, person);
         return person;
     }
 
     @Transactional
     public Address addAddressForPerson(int id, String addressName) {
-        var person = personDao.findById(id).get();
+        var person = personDao.findById(id).orElseThrow(NotFoundSource::new);
         var newAddress = addressDao.save(new Address(null, addressName, person));
         person.getAddresses().add(newAddress);
         return newAddress;
@@ -50,6 +51,7 @@ public class PersonService {
 
     @Transactional
     public void deleteAddress(int id) {
+        if (!addressDao.existsById(id)) throw new NotFoundSource();
         addressDao.deleteById(id);
     }
 }
